requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0
pandas>=2.0.0

# Model Inference
vllm>=0.2.0
ollama>=0.1.0
openai>=1.0.0  # for API compatibility

# Data Processing
datasets>=2.14.0
jieba>=0.42.1  # Chinese text processing
opencc-python-reimplemented>=0.1.7  # Traditional/Simplified Chinese conversion

# Machine Learning & Analysis
scikit-learn>=1.3.0
scipy>=1.11.0
umap-learn>=0.5.4  # for dimensionality reduction
hdbscan>=0.8.33  # for clustering

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
kaleido>=0.2.1  # for static image export

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0
python-dotenv>=1.0.0

# Testing & Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0

# Optional: For web interface
gradio>=4.0.0
streamlit>=1.28.0
